move subsriber trait back to mod
